Refactored URI handling
